Match aarch64 GNU-triples more general
The triplet is (currently) used to specify compiler flags which are
independent of the used c-libary and os.  Thus we ignore everything
except the architecture.
For safety reasons we still match against the triplet structure
but take into consideration that the triplet is sometimes a quadruple,
such as when using a "crosstool-ng" generated toolchain.